refactor(tools): remove unused file tools and reorganize imports
- Deleted the read-file, write-file, and match tools as they are no longer needed.
- Updated the index file to reflect these removals and reorganized imports for clarity.
- Moved perplexity-related functions to a new utils file for better structure.
